gvpr openFile: squash some -Wsign-compare warnings
GCC says:

  actions.c: In function ‘openFile’:
  actions.c:615:21: warning: comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’ [-Wsign-compare]
    615 |   for (idx = 3; idx < elementsof(ex->file); idx++)
        |                     ^

We know from context `idx` can never be negative.

Reported-by: GCC 11.4.0 -Wsign-compare

+++ lib/gvpr/actions.c
@@ -612,10 +612,10 @@ int openFile(Expr_t *ex, const char *fname, const char *mode) {
   int idx;
 
   /* find open index */
-  for (idx = 3; idx < elementsof(ex->file); idx++)
+  for (idx = 3; (size_t)idx < elementsof(ex->file); idx++)
     if (!ex->file[idx])
       break;
-  if (idx == elementsof(ex->file)) {
+  if ((size_t)idx == elementsof(ex->file)) {
     exerror("openF: no available descriptors");
     return -1;
   }
